+++ example/lib/main.dart
@@ -1,26 +1,27 @@
 import 'package:flutter_candies_demo_library/flutter_candies_demo_library.dart';
-import 'package:example/pages/no_route.dart';
 import 'package:extended_image/extended_image.dart';
 import 'package:extended_image_library/extended_image_library.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:image_picker_saver/image_picker_saver.dart';
-import "package:oktoast/oktoast.dart";
+import 'package:oktoast/oktoast.dart';
 import 'example_route.dart';
 import 'example_route_helper.dart';
 
 void main() => runApp(MyApp());
 
 class MyApp extends StatelessWidget {
-  final TuChongRepository listSourceRepository = TuChongRepository();
   MyApp() {
-    if (!kIsWeb) clearDiskCachedImages(duration: Duration(days: 7));
-    listSourceRepository.loadData().then((result) {
-      if (listSourceRepository.length > 0)
+    if (!kIsWeb) {
+      clearDiskCachedImages(duration: const Duration(days: 7));
+    }
+    listSourceRepository.loadData().then((bool result) {
+      if (listSourceRepository.isNotEmpty) {
         _imageTestUrl = listSourceRepository.first.imageUrl;
+      }
     });
   }
+  final TuChongRepository listSourceRepository = TuChongRepository();
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
@@ -31,26 +32,13 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      navigatorObservers: [
-        FFNavigatorObserver(routeChange: (name) {
-          //you can track page here
-          // print(name);
-        }, showStatusBarChange: (bool showStatusBar) {
-          // if (showStatusBar) {
-          //   SystemChrome.setEnabledSystemUIOverlays(SystemUiOverlay.values);
-          //   SystemChrome.setSystemUIOverlayStyle(SystemUiOverlayStyle.dark);
-          // } else {
-          //   SystemChrome.setEnabledSystemUIOverlays([]);
-          // }
-        })
-      ],
-      builder: (c, w) {
+      builder: (BuildContext c, Widget w) {
         ScreenUtil.init(width: 750, height: 1334, allowFontScaling: true);
         // ScreenUtil.instance =
         //     ScreenUtil(width: 750, height: 1334, allowFontScaling: true)
         //       ..init(c);
         if (!kIsWeb) {
-          final data = MediaQuery.of(c);
+          final MediaQueryData data = MediaQuery.of(c);
           return MediaQuery(
             data: data.copyWith(textScaleFactor: 1.0),
             child: w,
@@ -58,34 +46,52 @@ class MyApp extends StatelessWidget {
         }
         return w;
       },
-      initialRoute: "fluttercandies://mainpage",
+      initialRoute: 'fluttercandies://mainpage',
       onGenerateRoute: (RouteSettings settings) {
-        var routeResult =
-            getRouteResult(name: settings.name, arguments: settings.arguments);
+        String routeName = settings.name;
+        //when refresh web, route will as following
+        //   /
+        //   /fluttercandies:
+        //   /fluttercandies:/
+        //   /fluttercandies://mainpage
+
+        if (kIsWeb && routeName.startsWith('/')) {
+          routeName = routeName.replaceFirst('/', '');
+        }
+
+        final RouteResult routeResult = getRouteResult(
+            name: routeName,
+            arguments: settings.arguments as Map<String, dynamic>);
 
         if (routeResult.showStatusBar != null ||
             routeResult.routeName != null) {
           settings = FFRouteSettings(
               arguments: settings.arguments,
-              name: settings.name,
-              isInitialRoute: settings.isInitialRoute,
+              name: routeName,
               routeName: routeResult.routeName,
               showStatusBar: routeResult.showStatusBar);
         }
 
-        var page = routeResult.widget ?? NoRoute();
-        final platform = Theme.of(context).platform;
+        final Widget page = routeResult.widget ??
+            getRouteResult(
+                    name: 'fluttercandies://mainpage',
+                    arguments: settings.arguments as Map<String, dynamic>)
+                .widget;
+
+        final TargetPlatform platform = Theme.of(context).platform;
         switch (routeResult.pageRouteType) {
           case PageRouteType.material:
-            return MaterialPageRoute(settings: settings, builder: (c) => page);
+            return MaterialPageRoute<void>(
+                settings: settings, builder: (BuildContext c) => page);
           case PageRouteType.cupertino:
-            return CupertinoPageRoute(settings: settings, builder: (c) => page);
+            return CupertinoPageRoute<void>(
+                settings: settings, builder: (BuildContext c) => page);
           case PageRouteType.transparent:
             return platform == TargetPlatform.iOS
-                ? TransparentCupertinoPageRoute(
-                    settings: settings, builder: (c) => page)
-                : TransparentMaterialPageRoute(
-                    settings: settings, builder: (c) => page);
+                ? TransparentCupertinoPageRoute<void>(
+                    settings: settings, builder: (BuildContext c) => page)
+                : TransparentMaterialPageRoute<void>(
+                    settings: settings, builder: (BuildContext c) => page);
 //            return FFTransparentPageRoute(
 //                settings: settings,
 //                pageBuilder: (BuildContext context, Animation<double> animation,
@@ -93,8 +99,10 @@ class MyApp extends StatelessWidget {
 //                    page);
           default:
             return platform == TargetPlatform.iOS
-                ? CupertinoPageRoute(settings: settings, builder: (c) => page)
-                : MaterialPageRoute(settings: settings, builder: (c) => page);
+                ? CupertinoPageRoute<void>(
+                    settings: settings, builder: (BuildContext c) => page)
+                : MaterialPageRoute<void>(
+                    settings: settings, builder: (BuildContext c) => page);
         }
       },
     ));
@@ -103,12 +111,4 @@ class MyApp extends StatelessWidget {
 
 String _imageTestUrl;
 String get imageTestUrl =>
-    _imageTestUrl ?? "https://photo.tuchong.com/4870004/f/298584322.jpg";
-
-///save netwrok image to photo
-Future<bool> saveNetworkImageToPhoto(String url, {bool useCache: true}) async {
-  if (kIsWeb) return false;
-  var data = await getNetworkImageData(url, useCache: useCache);
-  var filePath = await ImagePickerSaver.saveFile(fileData: data);
-  return filePath != null && filePath != "";
-}
+    _imageTestUrl ?? 'https://photo.tuchong.com/4870004/f/298584322.jpg';