+++ CHANGELOG.md
@@ -1,8 +1,43 @@
+## [0.8.0]
+
+* Features:
+  Add call back [CanScrollPage] for ExtendedImageGesturePageView. 
+
+## [0.7.4]
+
+* Issues:
+  Fix ScrollPhysics is not working for ExtendedImageGesturePageView
+  
+## [0.7.3+1]
+
+* remove docs.
+
+## [0.7.3]
+
+* Improve:
+  fix build error on high flutter sdk(> 1.6.0)
+  fix analysiz_options
+
+## [0.7.2]
+
+* Features:
+  support loading progress for network
+  public HttpClient of ExtendedNetworkImageProvider
+  public ExtendedImageGestureState for SlideOffsetHandler/SlideEndHandler/SlideScaleHandler to get scale of image
+  
+## [0.7.1]
+
+* Improve:
+  scale parameter of method(handleDoubleTap) is support animationMinScale and animationMaxScale now.  
+
 ## [0.7.0]
 
 * Features:
   Support web.
-  Add clearMemoryCacheWhenDispose parameter that whether  clear memory cache when image is disposed.
+  Add clearMemoryCacheWhenDispose parameter that whether clear memory cache when image is disposed.
+* Issues:
+  Fix animationMinScale and animationMaxScale are not working for gif.  
+  Fix scale parameter of method(handleDoubleTap) is beyond minScale and maxScale.
 
 ## [0.6.9]
 
@@ -131,7 +166,7 @@
 
 ## [0.3.6]
 
-* handle loadfailed when re-addListener
+* handle load failed when re-addListener
 
 ## [0.3.4]
 
@@ -173,7 +208,7 @@
 ## [0.2.3]
 
 * add enableMemoryCache parameter, whether cache in PaintingBinding.instance.imageCache
-* add clearMemoryCacheIfFailed parameter, when failed to load image, whether clear memory cache,if ture, image will reload in next time.
+* add clearMemoryCacheIfFailed parameter, when failed to load image, whether clear memory cache,if true, image will reload in next time.
 * auto cancel network request is obsolete.
 
 ## [0.2.2]